+++ bin/alexpro100_lib.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 ###############################################################
 ### alexpro100 BASH LIBRARY
 ### Copyright (C) 2021 ALEXPRO100 (ktifhfl)
@@ -5,11 +6,13 @@
 ###############################################################
 shopt -s expand_aliases
 
-export ALEXPRO100_LIB_VERSION="0.2.4" ALEXPRO100_LIB_LOCATION="$(realpath ${BASH_SOURCE[0]})"
+ALEXPRO100_LIB_VERSION="0.2.4" 
+ALEXPRO100_LIB_LOCATION="$(realpath ${BASH_SOURCE[0]})"
+export ALEXPRO100_LIB_VERSION ALEXPRO100_LIB_LOCATION
 echo -e "ALEXPRO100 BASH LIBRARY $ALEXPRO100_LIB_VERSION"
 export TMP='' CHROOT_ACTIVE_MOUNTS=() CHROOT_CREATED=() ROOTFS_DIR_NO_FIX=0
 [[ -z "$ALEXPRO100_LIB_DEBUG" ]] && export ALEXPRO100_LIB_DEBUG=0
-alias AP100_DBG="[[ ! "$ALEXPRO100_LIB_DEBUG" == "1" ]] || "
+alias AP100_DBG='[[ ! $ALEXPRO100_LIB_DEBUG == 1 ]] || '
 
 # Colors for text.
 
@@ -57,13 +60,13 @@ function msg_print() {
   [[ -z $2 ]] && echo_help "Example: $FUNCNAME msg text" && return 1
   local TYPE=$1; shift
     case $TYPE in
-	  alert) echo -e "$Bold$White$On_Red$@${NC}";;
-	  err|error) echo -e "[${LRed}ERROR${NC}] $@";;
-	  warn|warning) echo -e "[${Orange}WARNING${NC}] $@";;
-	  note) echo -e "[${LBlue}NOTE${NC}] $@";;
-	  msg|meassage) echo -e "[${DGray}MSG${NC}] $@";;
-    debug) echo -e "[${LGreen}DEBUG${NC}] $@";;
-	  prgs|progress) echo -e "[${Orange}PROGRESS${NC}] $@";;
+	  alert) echo -e "$Bold$White$On_Red$*${NC}";;
+	  err|error) echo -e "[${LRed}ERROR${NC}] $*";;
+	  warn|warning) echo -e "[${Orange}WARNING${NC}] $*";;
+	  note) echo -e "[${LBlue}NOTE${NC}] $*";;
+	  msg|meassage) echo -e "[${DGray}MSG${NC}] $*";;
+    debug) echo -e "[${LGreen}DEBUG${NC}] $*";;
+	  prgs|progress) echo -e "[${Orange}PROGRESS${NC}] $*";;
 	  *);;
     esac
 }
@@ -108,10 +111,10 @@ function get_file_s() {
   [[ -z $2 ]] && echo_help "Example: $FUNCNAME file http://example.com/file" && return 1
   if command_exists wget; then
     AP100_DBG msg_print debug "Using wget."
-    wget -q -t 3 -O $1 $2 || return_err "Exit code $? while downloading $2!"
+    wget -q -t 3 -O "$1" "$2" || return_err "Exit code $? while downloading $2!"
   elif command_exists curl; then
     AP100_DBG msg_print debug "Using curl."
-    curl -s --retry 3 -f -o $1 $2 || return_err "Exit code $? while downloading $2!"
+    curl -s --retry 3 -f -o "$1" "$2" || return_err "Exit code $? while downloading $2!"
   else
     return_err "Niether wget nor curl are installed."
   fi
@@ -122,10 +125,10 @@ function check_url() {
   [[ -z $1 ]] && echo_help "Example: $FUNCNAME http://example.com/file" && return 1
   if command_exists wget; then
     AP100_DBG msg_print debug "Using wget."
-    wget -q --spider $1
+    wget -q --spider "$1"
   elif command_exists curl; then
     AP100_DBG msg_print debug "Using curl."
-    curl --head --fail $1 &>/dev/null
+    curl --head --fail "$1" &>/dev/null
   else
     return_err "Niether wget nor curl are installed."
   fi
@@ -136,14 +139,14 @@ function create_tmp_dir() {
   [[ -z $1 ]] && echo_help "Example: $FUNCNAME var_name" && return 1
   export $1="/tmp/.a_tmp_$RANDOM"
   AP100_DBG msg_print debug "Created tmp dir $1=${!1}."
-  mkdir -p ${!1} &>/dev/null
+  mkdir -p "${!1}" &>/dev/null
 }
 export -f create_tmp_dir
 
 #--- ROOTFS MOUNT: BEGIN
 function chroot_add_mount() {
   if [[ ! -e $3 ]]; then
-    [[ $1 == dir ]] && mkdir -p $3; [[ $1 == file ]] && touch $3
+    [[ $1 == dir ]] && mkdir -p "$3"; [[ $1 == file ]] && touch "$3"
     AP100_DBG msg_print debug "Created $1 $3."
     CHROOT_CREATED=("$3" "${CHROOT_CREATED[@]}")
   fi
@@ -177,7 +180,6 @@ function chroot_setup() {
 export -f chroot_setup
 
 function chroot_setup_light() {
-  CHROOT_POINTS=(proc sys dev dev/pts dev/shm run tmp)
   for mount_point in proc sys dev dev/pts dev/shm run tmp; do
     chroot_add_mount dir "/$mount_point" "$1/$mount_point" --bind
   done
@@ -190,11 +192,15 @@ export -f chroot_setup_light
 function chroot_teardown() {
   AP100_DBG msg_print debug "Running $FUNCNAME..."
   if (( ${#CHROOT_ACTIVE_MOUNTS[@]} )); then
-    AP100_DBG msg_print debug "Unmounting $(echo ${CHROOT_ACTIVE_MOUNTS[@]} | uniq)..."
-    umount --lazy $(echo ${CHROOT_ACTIVE_MOUNTS[@]} | uniq) || msg_print warning "Not 0 code exit!"
+    for name in "${CHROOT_ACTIVE_MOUNTS[@]}"; do
+      AP100_DBG msg_print debug "Unmounting $name..."
+      umount --lazy "$name" || msg_print warning "Not 0 code exit!"
+    done
     if [[ "$1" == "--remove-created" ]]; then
-      AP100_DBG msg_print debug "Removing $(echo ${CHROOT_CREATED[@]} | uniq)..."
-      rm -rf $(echo ${CHROOT_CREATED[@]} | uniq) || msg_print warning "Not 0 code exit!"
+      for name in "${CHROOT_CREATED[@]}"; do
+        AP100_DBG msg_print debug "Removing $name..."
+        rm -rf "$name" || msg_print warning "Not 0 code exit!"
+      done
     fi
   fi
   unset CHROOT_ACTIVE_MOUNTS CHROOT_CREATED
@@ -212,12 +218,12 @@ function chroot_rootfs() {
   local CHROOT_DIR="$2"; shift 2; [[ -z $CHROOT_COMMAND ]] && local CHROOT_COMMAND=chroot
   if [[ $ROOTFS_DIR_NO_FIX == 0 ]] && ! mountpoint -q "$CHROOT_DIR"; then
     msg_print warning "Not mounted directory. Bypassing..."
-    chroot_add_mount dir $CHROOT_DIR $CHROOT_DIR --bind
+    chroot_add_mount dir "$CHROOT_DIR" "$CHROOT_DIR" --bind
   fi
-  chroot_setup"$ADD_COMMAND" $CHROOT_DIR
+  chroot_setup"$ADD_COMMAND" "$CHROOT_DIR"
   AP100_DBG msg_print debug "Running chroot..."
-  unshare --fork $CHROOT_COMMAND $CHROOT_DIR "$@" || msg_print warning "Not 0 code exit!"
-  chroot_teardown
+  unshare --fork $CHROOT_COMMAND "$CHROOT_DIR" "$@" || msg_print warning "Not 0 code exit!"
+  chroot_teardown ""
 }
 export -f chroot_rootfs
 
@@ -230,7 +236,10 @@ function parse_arch() {
     armhf|armv6h) export alpine_arch=armhf debian_arch=armhf arch_arch=armv6h void_arch=armv6l qemu_arch=arm ;;
     arm|armel) export alpine_arch=armhf debian_arch=armel arch_arch=arm void_arch=armv6l qemu_arch=arm;;
     # TODO: Add and fix another arches (old arm, mips).
-    *) export alpine_arch=$(uname -m) debian_arch=$(uname -m) arch_arch=$(uname -m) void_arch=armv6l qemu_arch=$(uname -m);;
+    *) 
+    alpine_arch="$(uname -m)" debian_arch="$(uname -m)" arch_arch="$(uname -m)" void_arch="armv6l" qemu_arch="$(uname -m)"
+    export alpine_arch debian_arch arch_arch void_arch qemu_arch
+    ;;
   esac
   AP100_DBG msg_print debug "Exported alpine_arch=$alpine_arch debian_arch=$debian_arch arch_arch=$arch_arch void_arch=$void_arch qemu_arch=$qemu_arch."
 }
@@ -240,17 +249,17 @@ function qemu_chroot() {
   [[ -z $3 ]] && echo_help "Example: $FUNCNAME aarch64 /mnt/mnt ash" && return 1
   [[ -z $QEMU_STATIC_BIN_DIR ]] && local QEMU_STATIC_BIN_DIR="/usr/bin"
   if [[ "$1" == "check" ]]; then
-    parse_arch $2
+    parse_arch "$2"
   else
-    parse_arch $1; shift
+    parse_arch "$1"; shift
   fi
   if [[ -f $QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static ]]; then
     AP100_DBG msg_print debug "Found $QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static."
     [[ ! "$1" == "check" ]] || return 0
-    cp $QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static $1/usr/bin/qemu-$qemu_arch-static
+    cp "$QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static" "$1/usr/bin/qemu-$qemu_arch-static"
     local qemu_dir=$1; shift
-	  chroot_rootfs main $qemu_dir qemu-$qemu_arch-static "$@"
-    rm -rf $1/usr/bin/qemu-$qemu_arch-static
+	  chroot_rootfs main "$qemu_dir" "qemu-$qemu_arch-static" "$@"
+    rm -rf "$1/usr/bin/qemu-$qemu_arch-static"
   else
     return_err "File $QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static not found! Check qemu-static package."
   fi
@@ -259,10 +268,10 @@ export -f qemu_chroot
 
 function qemu_run_bin() {
   [[ -z $2 ]] && echo_help "Example: $FUNCNAME aarch64 /bin/ash" && return 1
-  [[ -z $QEMU_STATIC_BIN_DIR ]] && local QEMU_STATIC_BIN_DIR="/usr/bin"
-  if [[ -f $QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static ]]; then
+  [[ -z "$QEMU_STATIC_BIN_DIR" ]] && local QEMU_STATIC_BIN_DIR="/usr/bin"
+  if [[ -f "$QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static" ]]; then
     AP100_DBG msg_print debug "Found $QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static."
-    shift; $QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static $@
+    shift; "$QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static" "$@"
   else
     return_err "File $QEMU_STATIC_BIN_DIR/qemu-$qemu_arch-static not found! Check qemu-static package."
   fi
@@ -271,7 +280,8 @@ export -f qemu_run_bin
 
 function genfstab_light() {
   [[ -z $1 ]] && echo_help "Example: $FUNCNAME /mnt/mnt" && return 1
-  local root=$(realpath -mL "$1")
+  local root
+  root=$(realpath -mL "$1")
   declare -A pseudofs_types=([anon_inodefs]=1 [autofs]=1 [bdev]=1 [bpf]=1 [binfmt_misc]=1 [cgroup]=1 [cgroup2]=1 [configfs]=1 [cpuset]=1 [debugfs]=1
   [devfs]=1 [devpts]=1 [devtmpfs]=1 [dlmfs]=1 [efivarfs]=1 [fuse.gvfsd-fuse]=1 [fuse.gvfs-fuse-daemon]=1 [fusectl]=1 [gvfsd-fuse]=1 [hugetlbfs]=1 [mqueue]=1 [nfsd]=1 [none]=1 [pipefs]=1
   [proc]=1 [pstore]=1 [ramfs]=1 [rootfs]=1 [rpc_pipefs]=1 [securityfs]=1 [sockfs]=1 [spufs]=1 [sysfs]=1 [tracefs]=1 [tmpfs]=1)
@@ -285,14 +295,14 @@ function genfstab_light() {
         echo "#Warning! BTRFS was not tested!"
         opts+=,subvol=${fsroot#/}
       else
-        [[ $(findmnt -funcevo TARGET "$src")$fsroot != $target ]] && continue
+        [[ $(findmnt -funcevo TARGET "$src")$fsroot != "$target" ]] && continue
       fi
     fi
     dump=0 pass=0
     (( fsck_types["$fstype"] )) && pass=2
     findmnt "$src" "$root" >/dev/null && pass=1
     [[ $fstype == fuseblk ]] && fstype=$(lsblk -no FSTYPE "$src")
-    echo -ne "\n#$src"; label=$(lsblk -rno LABEL "$src" 2>/dev/null); [[ ! -z $label ]] && echo -ne " LABEL=$label"
+    echo -ne "\n#$src"; label=$(lsblk -rno LABEL "$src" 2>/dev/null); [[ -n $label ]] && echo -ne " LABEL=$label"
     echo -ne "\nUUID=$(lsblk -rno UUID "$src" 2>/dev/null)\t/${target#/}\t$fstype\t$opts\t$dump $pass\n"
   done
 }